update to remove extra asset image indirection
