+++ packages/flutter/lib/src/painting/_network_image_web.dart
@@ -39,9 +39,11 @@ class NetworkImage extends image_provider.ImageProvider<image_provider.NetworkIm
     return MultiFrameImageStreamCompleter(
       codec: _loadAsync(key),
       scale: key.scale,
-      informationCollector: () sync* {
-        yield DiagnosticsProperty<image_provider.ImageProvider>('Image provider', this);
-        yield DiagnosticsProperty<NetworkImage>('Image key', key);
+      informationCollector: () {
+        return <DiagnosticsNode>[
+          DiagnosticsProperty<image_provider.ImageProvider>('Image provider', this),
+          DiagnosticsProperty<NetworkImage>('Image key', key),
+        ];
       },
     );
   }
